requirements.txt
requests==2.31.0
yfinance==0.2.36
ratelimit==2.2.1
pandas==2.1.4
python-dotenv==1.0.0